Move module version related types
From sandboxTypes.ts to moduleVersion.ts
